+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Changed
 
 - [core] Dropped BSON dependency
+- [ui] Logout user when in ephemeral mode
 
 ### Fixed